requirements.txt
pandas
sqlalchemy
psycopg2-binary
requests
os
python-dotenv